django-auth/requirements.txt
asgiref==3.5.2
boto3==1.24.66
botocore==1.27.66
certifi==2022.6.15
charset-normalizer==2.1.1
dj-database-url==1.0.0
Django==3.2.15
django-cors-headers==3.13.0
django-dotenv==1.4.2
django-extensions==3.1.5
django-storages==1.13.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
djongo==1.3.6
dnspython==2.2.1
gunicorn==20.1.0
idna==3.3
jmespath==1.0.1
psycopg2-binary==2.9.3
PyJWT==2.4.0
pymongo==4.2.0
python-dateutil==2.8.2
pytz==2022.2.1
requests==2.28.1
s3transfer==0.6.0
six==1.16.0
sqlparse==0.2.4
urllib3==1.26.12
whitenoise==6.2.0
